Natural Language Processing task
The aim of this study is to use various machine learning models, including logistic regression, decision trees, random forests, and support vector machines, to categorize Reddit posts from four specific subreddits: Trump, Ford, Obama, and Musk. Raw text data is preprocessed through techniques like text cleaning, tokenization, stop word
removal, and stemming before being converted into numerical values. To determine the best models for the final test data on Kaggle, the K-Fold cross-validation technique is utilized on the validation and training data. Additionally, the models are compared to a Naive Bayes model built from scratch in terms of accuracy and run-time based on the validation data. The findings reveal that the MLP model is the most accurate with a %97.59 in the Kaggle competition.